+++ crates/teloxide/src/dispatching.rs
@@ -226,7 +226,9 @@ mod handler_ext;
 mod tracing;
 
 pub use crate::utils::shutdown_token::{IdleShutdownError, ShutdownToken};
-pub use dispatcher::{Dispatcher, DispatcherBuilder, UpdateHandler};
+pub use dispatcher::{
+    Dispatcher, DispatcherBuilder, UpdateHandler, WorkerDispatchError, WorkerTermination,
+};
 pub use distribution::DefaultKey;
 pub use filter_ext::{MessageFilterExt, UpdateFilterExt};
 pub use handler_description::DpHandlerDescription;